backend/requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==4.2.14
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-gunicorn==0.1.1
django-rest-framework==0.1.0
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-api-key==3.0.0
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.1
gunicorn==22.0.0
idna==3.7
packaging==24.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
